#298 refactor code
